+++ src/compas_timber/fabrication/__init__.py
@@ -26,6 +26,8 @@ from .btlx import AlignmentType
 from .btlx import MachiningLimits
 from .btlx import StepShapeType
 from .btlx import BTLxFromGeometryDefinition
+from .btlx import Contour
+from .btlx import DualContour
 
 __all__ = [
     "BTLxWriter",
@@ -56,4 +58,6 @@ __all__ = [
     "MachiningLimits",
     "StepShapeType",
     "BTLxFromGeometryDefinition",
+    "Contour",
+    "DualContour",
 ]