+++ src/compas_timber/fabrication/btlx.py
@@ -251,7 +251,7 @@ class BTLxWriter(object):
         return processing_element
 
     def _element_from_complex_param(self, param):
-        serializer = self.SERIALIZERS.get(type(param), None)
+        serializer = self.SERIALIZERS.get(type(param).__name__, None)
         if not serializer:
             raise ValueError("No serializer found for type: {}".format(type(param)))
         return serializer(param)
@@ -486,13 +486,12 @@ def contour_to_xml(contour):
         The element which represents the contour.
 
     """
+
     root = ET.Element("Contour")
     if contour.depth:
         root.set("Depth", "{:.{prec}f}".format(contour.depth, prec=BTLxWriter.POINT_PRECISION))
     if contour.depth_bounded:
         root.set("DepthBounded", "yes" if contour.depth_bounded else "no")
-    if contour.inclination:
-        root.set("Inclination", "{:.{prec}f}".format(contour.inclination, prec=BTLxWriter.ANGLE_PRECISION))
 
     start = contour.polyline[0]
     start_point = ET.SubElement(root, "StartPoint")
@@ -500,12 +499,56 @@ def contour_to_xml(contour):
     start_point.set("Y", "{:.{prec}f}".format(start.y, prec=BTLxWriter.POINT_PRECISION))
     start_point.set("Z", "{:.{prec}f}".format(start.z, prec=BTLxWriter.POINT_PRECISION))
 
-    for point in contour.polyline[1:]:
-        line = ET.SubElement(root, "Line")
-        end_point = ET.SubElement(line, "EndPoint")
-        end_point.set("X", "{:.{prec}f}".format(point[0], prec=BTLxWriter.POINT_PRECISION))
-        end_point.set("Y", "{:.{prec}f}".format(point[1], prec=BTLxWriter.POINT_PRECISION))
-        end_point.set("Z", "{:.{prec}f}".format(point[2], prec=BTLxWriter.POINT_PRECISION))
+    if len(contour.inclination) == 1:  # single Inclination for all segments
+        root.set("Inclination", "{:.{prec}f}".format(contour.inclination[0], prec=BTLxWriter.ANGLE_PRECISION))
+        for point in contour.polyline[1:]:
+            line = ET.SubElement(root, "Line")
+            end_point = ET.SubElement(line, "EndPoint")
+            end_point.set("X", "{:.{prec}f}".format(point[0], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Y", "{:.{prec}f}".format(point[1], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Z", "{:.{prec}f}".format(point[2], prec=BTLxWriter.POINT_PRECISION))
+
+    else:  # one Inclination value per segment
+        for point, inc in zip(contour.polyline[1:], contour.inclination):
+            line = ET.SubElement(root, "Line", {"Inclination": "{:.{prec}f}".format(inc, prec=BTLxWriter.ANGLE_PRECISION)})
+            end_point = ET.SubElement(line, "EndPoint")
+            end_point.set("X", "{:.{prec}f}".format(point[0], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Y", "{:.{prec}f}".format(point[1], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Z", "{:.{prec}f}".format(point[2], prec=BTLxWriter.POINT_PRECISION))
+
+    return root
+
+
+def dual_contour_to_xml(contour):
+    """Converts a contour to an XML element.
+
+    Parameters
+    ----------
+    contour : :class:`DualContour`
+        The DualContour to be converted.
+
+    Returns
+    -------
+    :class:`~xml.etree.ElementTree.Element`
+        The element which represents the contour.
+
+    """
+    root = ET.Element("DualContour")
+    principal_contour = ET.SubElement(root, "PrincipalContour")
+    associated_contour = ET.SubElement(root, "AssociatedContour")
+    for polyline, et_contour in zip([contour.principal_contour, contour.associated_contour], [principal_contour, associated_contour]):
+        start = polyline[0]
+        start_point = ET.SubElement(et_contour, "StartPoint")
+        start_point.set("X", "{:.{prec}f}".format(start.x, prec=BTLxWriter.POINT_PRECISION))
+        start_point.set("Y", "{:.{prec}f}".format(start.y, prec=BTLxWriter.POINT_PRECISION))
+        start_point.set("Z", "{:.{prec}f}".format(start.z, prec=BTLxWriter.POINT_PRECISION))
+
+        for point in polyline[1:]:
+            line = ET.SubElement(et_contour, "Line")
+            end_point = ET.SubElement(line, "EndPoint")
+            end_point.set("X", "{:.{prec}f}".format(point[0], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Y", "{:.{prec}f}".format(point[1], prec=BTLxWriter.POINT_PRECISION))
+            end_point.set("Z", "{:.{prec}f}".format(point[2], prec=BTLxWriter.POINT_PRECISION))
     return root
 
 
@@ -757,8 +800,10 @@ class AlignmentType(object):
     CENTER = "center"
 
 
-class Contour(object):
+class Contour(Data):
     """Represens the generic contour for specific free contour processings.
+
+    TODO: add point attributes for other types like NailContour
 
     Parameters
     ----------
@@ -770,17 +815,50 @@ class Contour(object):
         The inclination of the contour.
     polyline : :class:`compas.geometry.Polyline`
         The polyline of the contour.
-    TODO: add point attributes for other types like NailContour
     """
 
-    def __init__(self, polyline, depth=None, depth_bounded=None, inclination=None):
+    def __init__(self, polyline, depth=None, depth_bounded=True, inclination=None):
+        super(Contour, self).__init__()
+        self.polyline = polyline
         self.depth = depth
         self.depth_bounded = depth_bounded
         self.inclination = inclination
-        self.polyline = polyline
+
+    @property
+    def __data__(self):
+        return {"polyline": self.polyline, "depth": self.depth, "depth_bounded": self.depth_bounded, "inclination": self.inclination}
 
 
-BTLxWriter.register_type_serializer(Contour, contour_to_xml)
+BTLxWriter.register_type_serializer(Contour.__name__, contour_to_xml)
+
+
+class DualContour(Data):
+    """Represens the generic contour for specific free contour processings.
+
+    TODO: add point attributes for other types like NailContour
+
+    Parameters
+    ----------
+    principal_contour : :class:`compas.geometry.Polyline`
+        The principal contour of the dual contour.
+    associated_contour : :class:`compas.geometry.Polyline`
+        The associated contour of the dual contour. Must have same number of segments as `principal_contour`.
+    depth_bounded : bool
+        If True, the depth is bounded.
+    """
+
+    def __init__(self, principal_contour, associated_contour, depth_bounded=None):
+        super(DualContour, self).__init__()
+        self.principal_contour = principal_contour
+        self.associated_contour = associated_contour
+        self.depth_bounded = depth_bounded
+
+    @property
+    def __data__(self):
+        return {"principal_contour": self.principal_contour, "associated_contour": self.associated_contour, "depth_bounded": self.depth_bounded}
+
+
+BTLxWriter.register_type_serializer(DualContour.__name__, dual_contour_to_xml)
 
 
 class BTLxFromGeometryDefinition(Data):